Added admin functions and fixed bugs

+++ poli-meets-core/src/main/java/poli/meets/coreservice/web/rest/TeacherResource.java
@@ -1,10 +1,10 @@
 package poli.meets.coreservice.web.rest;
 
 import lombok.AllArgsConstructor;
+import poli.meets.coreservice.client.AuthClient;
 import poli.meets.coreservice.service.TeacherService;
 import poli.meets.coreservice.service.dto.TeacherDTO;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,6 +13,7 @@ import java.net.URISyntaxException;
 import java.util.List;
 
 import lombok.extern.slf4j.Slf4j;
+import poli.meets.coreservice.web.rest.errors.ForbiddenException;
 
 /**
  * REST controller for managing {@link poli.meets.coreservice.domain.Teacher}.
@@ -25,62 +26,50 @@ public class TeacherResource {
 
     private final TeacherService teacherService;
 
-    /**
-     * {@code POST  /teachers} : Create a new teacher.
-     *
-     * @param teacherDTO the teacherDTO to create.
-     * @return the {@link ResponseEntity} with status {@code 201 (Created)} and with body the new teacherDTO, or with status {@code 400 (Bad Request)} if the teacher has already an ID.
-     * @throws URISyntaxException if the Location URI syntax is incorrect.
-     */
+    private final AuthClient authClient;
+
     @PostMapping("/teachers")
-    public ResponseEntity<TeacherDTO> createTeacher(@RequestBody TeacherDTO teacherDTO) throws URISyntaxException {
-        log.debug("REST request to save Teacher : {}", teacherDTO);
+    public ResponseEntity<TeacherDTO> createTeacher(
+            @RequestHeader("Authorization") String token,
+            @RequestBody TeacherDTO teacherDTO) throws URISyntaxException {
+        ResponseEntity<Boolean> isAdmin = authClient.isCurrentUserAdmin(token);
+        if (isAdmin == null || isAdmin.getBody() == null || !isAdmin.getBody()) {
+            throw new ForbiddenException();
+        }
 
         TeacherDTO result = teacherService.save(teacherDTO);
         return ResponseEntity.created(new URI("/api/teachers/" + result.getId()))
             .body(result);
     }
 
-    /**
-     * {@code PUT  /teachers} : Updates an existing teacher.
-     *
-     * @param teacherDTO the teacherDTO to update.
-     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and with body the updated teacherDTO,
-     * or with status {@code 400 (Bad Request)} if the teacherDTO is not valid,
-     * or with status {@code 500 (Internal Server Error)} if the teacherDTO couldn't be updated.
-     * @throws URISyntaxException if the Location URI syntax is incorrect.
-     */
     @PutMapping("/teachers")
-    public ResponseEntity<TeacherDTO> updateTeacher(@RequestBody TeacherDTO teacherDTO) throws URISyntaxException {
-        log.debug("REST request to update Teacher : {}", teacherDTO);
+    public ResponseEntity<TeacherDTO> updateTeacher(
+            @RequestHeader("Authorization") String token,
+            @RequestBody TeacherDTO teacherDTO) {
+        ResponseEntity<Boolean> isAdmin = authClient.isCurrentUserAdmin(token);
+        if (isAdmin == null || isAdmin.getBody() == null || !isAdmin.getBody()) {
+            throw new ForbiddenException();
+        }
 
         TeacherDTO result = teacherService.save(teacherDTO);
         return ResponseEntity.ok()
             .body(result);
     }
 
-    /**
-     * {@code GET  /teachers} : get all the teachers.
-     *
-     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and the list of teachers in body.
-     */
     @GetMapping("/teachers")
     public ResponseEntity<List<TeacherDTO>> getAllTeachers() {
-        log.debug("REST request to get all Teachers");
         return ResponseEntity.ok(teacherService.findAll());
     }
 
-
-
-    /**
-     * {@code DELETE  /teachers/:id} : delete the "id" teacher.
-     *
-     * @param id the id of the teacherDTO to delete.
-     * @return the {@link ResponseEntity} with status {@code 204 (NO_CONTENT)}.
-     */
     @DeleteMapping("/teachers/{id}")
-    public ResponseEntity<Void> deleteTeacher(@PathVariable Long id) {
-        log.debug("REST request to delete Teacher : {}", id);
+    public ResponseEntity<Void> deleteTeacher(
+            @RequestHeader("Authorization") String token,
+            @PathVariable Long id) {
+        ResponseEntity<Boolean> isAdmin = authClient.isCurrentUserAdmin(token);
+        if (isAdmin == null || isAdmin.getBody() == null || !isAdmin.getBody()) {
+            throw new ForbiddenException();
+        }
+
         teacherService.delete(id);
         return ResponseEntity.noContent().build();
     }
